Getaway from Microsoft madness with types

+++ WinDiskWipe/Main.cpp
@@ -4,12 +4,12 @@
 
 namespace Help
 {
-    std::wstring Win32ErrorToString(const DWORD error)
+    std::wstring Win32ErrorToString(const uint32_t error)
     {
-        constexpr DWORD flags = FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS;
-        constexpr DWORD langId = MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT);
+        constexpr uint32_t flags = FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS;
+        constexpr uint32_t langId = MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT);
         wchar_t buffer[WDW::KiloByte] = { 0 };
-        const DWORD size = FormatMessage(flags, nullptr, error, langId, buffer, WDW::KiloByte, nullptr);
+        const uint32_t size = FormatMessage(flags, nullptr, error, langId, buffer, WDW::KiloByte, nullptr);
         return std::wstring(buffer, size);
     }
 
@@ -19,7 +19,7 @@ namespace Help
         std::wcout << L"Example: \\\\.\\PHYSICALDRIVE1" << std::endl;
         std::wcout << std::endl << "Here are your drives:" << std::endl << std::endl;
 
-        DWORD error = static_cast<DWORD>(_wsystem(L"wmic diskdrive list brief"));
+        int error = _wsystem(L"wmic diskdrive list brief");
 
         if (error != 0)
         {
@@ -38,35 +38,35 @@ int wmain(int argc, wchar_t* argv[])
 
     std::wcout << L"Hello. Going to wipe your drive: " << argv[1] << std::endl;
 
-    constexpr DWORD desiredAccess = GENERIC_READ | GENERIC_WRITE;
-    constexpr DWORD shareMode = FILE_SHARE_READ | FILE_SHARE_WRITE;
+    constexpr uint32_t desiredAccess = GENERIC_READ | GENERIC_WRITE;
+    constexpr uint32_t shareMode = FILE_SHARE_READ | FILE_SHARE_WRITE;
     const WDW::AutoHandle hdd = CreateFile(argv[1], desiredAccess, shareMode, nullptr, OPEN_EXISTING, 0, nullptr);
 
     if (!hdd.IsValid())
     {
-        const DWORD error = GetLastError();
+        const uint32_t error = GetLastError();
         std::wcerr << L"Could not open drive: " << error << L" / " << Help::Win32ErrorToString(error) << std::endl;
         return error;
     }
 
-    UINT64 bytesLeft = DiskSize(hdd);
+    uint64_t bytesLeft = DiskSize(hdd);
 
     if (!bytesLeft)
     {
-        const DWORD error = GetLastError();
+        const uint32_t error = GetLastError();
         std::wcerr << L"Failed to get disk geometry: " << error << L" / " << Help::Win32ErrorToString(error) << std::endl;
         return error;
     }
 
     std::wcout << L"Bytes to wipe: " << bytesLeft << std::endl;
 
-    UINT64 writtenBytesTotal = 0;
-    DWORD error = ERROR_SUCCESS;
+    uint64_t writtenBytesTotal = 0;
+    uint32_t error = ERROR_SUCCESS;
     const Timer<std::chrono::seconds> stopWatch;
 
-    const auto progress = [](UINT64 bytes, std::chrono::seconds time) -> void
+    const auto progress = [](uint64_t bytes, std::chrono::seconds time) -> void
     {
-        const UINT64 megabytes = bytes / WDW::MegaByte;
+        const uint64_t megabytes = bytes / WDW::MegaByte;
         std::wcout << megabytes << L" megabytes written. Speed " << megabytes / time.count() << " MB/s" << std::endl;
     };
